# docs/openapi.yaml
openapi: 3.1.0
info:
  contact:
    name: 筑享开发团队
  description: 筑享移动端和管理端后端接口。除明确标注为公开接口外，受保护接口需在 Authorization 请求头中携带 Bearer 访问令牌。
  license:
    name: 内部使用
  title: 筑享租房平台接口文档
  version: v1
servers:
- description: Generated server url
  url: http://localhost:18080/api
...
paths:
...
  /rent-orders:
    post:
      description: 按房源、起租日期、租期、付款方式和入住人数创建订单并锁定初始费用。
      operationId: createOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRentOrderRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseRentOrderResponse"
          description: 请求成功
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求参数错误或业务校验失败
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 未登录、访问令牌缺失或已失效
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 当前用户无权执行该操作
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求的业务资源不存在
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 资源状态冲突，无法完成当前操作
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 服务器内部错误
      security:
      - bearerAuth: []
      summary: 创建租房订单
      tags:
      - 租房订单
  /rent-orders/{orderId}:
    get:
      description: 查询当前用户的租房订单、费用、实名、合同和支付状态。
      operationId: getOrderDetail
      parameters:
      - description: 租房订单 ID
        example: order_001
        in: path
        name: orderId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseRentOrderResponse"
          description: 请求成功
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求参数错误或业务校验失败
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 未登录、访问令牌缺失或已失效
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 当前用户无权执行该操作
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求的业务资源不存在
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 资源状态冲突，无法完成当前操作
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 服务器内部错误
      security:
      - bearerAuth: []
      summary: 获取订单详情
      tags:
      - 租房订单
  /rent-orders/{orderId}/confirm-contract:
    post:
      description: 确认当前订单的合同内容，确认后进入待支付阶段。
      operationId: confirmContract
      parameters:
      - description: 租房订单 ID
        example: order_001
        in: path
        name: orderId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseRentOrderResponse"
          description: 请求成功
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求参数错误或业务校验失败
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 未登录、访问令牌缺失或已失效
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 当前用户无权执行该操作
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求的业务资源不存在
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 资源状态冲突，无法完成当前操作
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 服务器内部错误
      security:
      - bearerAuth: []
      summary: 确认租赁合同
      tags:
      - 租房订单
  /rent-orders/{orderId}/contract-preview:
    get:
      description: 根据订单和实名认证信息生成合同预览数据。
      operationId: getContractPreview
      parameters:
      - description: 租房订单 ID
        example: order_001
        in: path
        name: orderId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseContractPreviewResponse"
          description: 请求成功
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求参数错误或业务校验失败
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 未登录、访问令牌缺失或已失效
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 当前用户无权执行该操作
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求的业务资源不存在
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 资源状态冲突，无法完成当前操作
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 服务器内部错误
      security:
      - bearerAuth: []
      summary: 预览租赁合同
      tags:
      - 租房订单
  /rent-orders/{orderId}/pay:
    post:
      description: 按支付方式和支付渠道发起付款；当前实现可使用 mock 渠道。
      operationId: pay
      parameters:
      - description: 租房订单 ID
        example: order_001
        in: path
        name: orderId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseRentOrderResponse"
          description: 请求成功
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求参数错误或业务校验失败
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 未登录、访问令牌缺失或已失效
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 当前用户无权执行该操作
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求的业务资源不存在
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 资源状态冲突，无法完成当前操作
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 服务器内部错误
      security:
      - bearerAuth: []
      summary: 支付租房订单
      tags:
      - 租房订单
  /rent-orders/{orderId}/payment-info:
    get:
      description: 返回订单首期应付金额、租金、押金、服务费和付款周期。
      operationId: getPaymentInfo
      parameters:
      - description: 租房订单 ID
        example: order_001
        in: path
        name: orderId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponsePaymentInfoResponse"
          description: 请求成功
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求参数错误或业务校验失败
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 未登录、访问令牌缺失或已失效
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 当前用户无权执行该操作
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求的业务资源不存在
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 资源状态冲突，无法完成当前操作
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 服务器内部错误
      security:
      - bearerAuth: []
      summary: 获取支付信息
      tags:
      - 租房订单
  /rent-orders/{orderId}/real-name:
    post:
      description: 为订单提交租客姓名、手机号、身份证号以及身份证正反面图片地址。
      operationId: submitRealName
      parameters:
      - description: 租房订单 ID
        example: order_001
        in: path
        name: orderId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RealNameRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseRentOrderResponse"
          description: 请求成功
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求参数错误或业务校验失败
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 未登录、访问令牌缺失或已失效
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 当前用户无权执行该操作
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求的业务资源不存在
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 资源状态冲突，无法完成当前操作
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 服务器内部错误
      security:
      - bearerAuth: []
      summary: 提交实名认证
      tags:
      - 租房订单
  /rent-orders/{orderId}/sign:
    post:
      description: 在合同确认并支付后完成签约，生成后续租约数据。
      operationId: sign
      parameters:
      - description: 租房订单 ID
        example: order_001
        in: path
        name: orderId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseRentOrderResponse"
          description: 请求成功
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求参数错误或业务校验失败
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 未登录、访问令牌缺失或已失效
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 当前用户无权执行该操作
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求的业务资源不存在
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 资源状态冲突，无法完成当前操作
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 服务器内部错误
      security:
      - bearerAuth: []
      summary: 完成订单签约
      tags:
      - 租房订单
components:
  schemas:
...
    ApiResponseContractPreviewResponse:
      type: object
      description: 统一接口响应
      properties:
        code:
          type: integer
          format: int32
          description: 业务状态码，与 HTTP 状态码保持一致
          example: 200
        data:
          $ref: "#/components/schemas/ContractPreviewResponse"
          description: 响应数据；无返回数据时为 null
        message:
          type: string
          description: 处理结果说明
          example: success
...
    ApiResponsePaymentInfoResponse:
      type: object
      description: 统一接口响应
      properties:
        code:
          type: integer
          format: int32
          description: 业务状态码，与 HTTP 状态码保持一致
          example: 200
        data:
          $ref: "#/components/schemas/PaymentInfoResponse"
          description: 响应数据；无返回数据时为 null
        message:
          type: string
          description: 处理结果说明
          example: success
    ApiResponseRentOrderResponse:
      type: object
      description: 统一接口响应
      properties:
        code:
          type: integer
          format: int32
          description: 业务状态码，与 HTTP 状态码保持一致
          example: 200
        data:
          $ref: "#/components/schemas/RentOrderResponse"
          description: 响应数据；无返回数据时为 null
        message:
          type: string
          description: 处理结果说明
          example: success
...
    ApiResponseVoid:
      type: object
      description: 统一接口响应
      properties:
        code:
          type: integer
          format: int32
          description: 业务状态码，与 HTTP 状态码保持一致
          example: 200
        data:
          description: 响应数据；无返回数据时为 null
        message:
          type: string
          description: 处理结果说明
          example: success
...
    ContractPreviewResponse:
      type: object
      description: 租赁合同预览
      properties:
        contractNo:
          type: string
          description: 合同编号
        deposit:
          type: integer
          format: int32
          description: 押金，单位元
        endDate:
          type: string
          format: date
          description: 结束日期
        firstPaymentAmount:
          type: integer
          format: int32
          description: 首期应付金额，单位元
        houseAddress:
          type: string
          description: 房源详细地址
        houseName:
          type: string
          description: 房源名称
        leaseMonths:
          type: integer
          format: int32
          description: 租期月数
        monthlyRent:
          type: integer
          format: int32
          description: 月租金，单位元
        paymentMonths:
          type: integer
          format: int32
          description: 每期支付月数
        roomName:
          type: string
          description: 房间名称
        serviceFee:
          type: integer
          format: int32
          description: 服务费，单位元
        startDate:
          type: string
          format: date
          description: 开始日期
        status:
          type: string
          description: 业务状态
        tenantIdCard:
          type: string
          description: 租客身份证号
        tenantName:
          type: string
          description: 租客姓名
        tenantPhone:
          type: string
          description: 租客联系电话
...
    CreateRentOrderRequest:
      type: object
      description: 租房订单创建请求
      properties:
        houseId:
          type: string
          description: 房源 ID
          example: house_001
          minLength: 1
        leaseMonths:
          type: integer
          format: int32
          description: 租期月数，范围 1-120
          example: 12
          maximum: 120
          minimum: 1
        paymentMethod:
          type: string
          description: 租金付款方式，如押一付一、押一付三
          example: 押一付三
          minLength: 1
        startDate:
          type: string
          format: date
          description: 起租日期，不能早于今天
          example: 2026-07-01
        tenantCount:
          type: integer
          format: int32
          description: 租住人数，范围 1-20
          example: 2
          maximum: 20
          minimum: 1
      required:
      - houseId
      - leaseMonths
      - paymentMethod
      - startDate
      - tenantCount
...
    PayRequest:
      type: object
      description: 租房订单支付请求
      properties:
        paymentChannel:
          type: string
          description: 支付渠道；不传时默认为 mock
          example: mock
        paymentMethod:
          type: string
          description: 订单付款方式，需与订单约定一致
          example: 押一付三
          minLength: 1
      required:
      - paymentMethod
    PaymentInfoResponse:
      type: object
      description: 订单支付信息
      properties:
        deposit:
          type: integer
          format: int32
          description: 押金，单位元
        firstPaymentAmount:
          type: integer
          format: int32
          description: 首期应付金额，单位元
        monthlyRent:
          type: integer
          format: int32
          description: 月租金，单位元
        orderId:
          type: string
          description: 租房订单 ID
        paymentMethod:
          type: string
          description: 租金付款方式
        paymentMonths:
          type: integer
          format: int32
          description: 每期支付月数
        serviceFee:
          type: integer
          format: int32
          description: 服务费，单位元
        status:
          type: string
          description: 业务状态
...
    RealNameRequest:
      type: object
      description: 订单实名认证请求
      properties:
        idCardBackUrl:
          type: string
          description: 身份证国徽面图片 URL
          example: /api/uploads/id-card-back.jpg
          minLength: 1
        idCardFrontUrl:
          type: string
          description: 身份证人像面图片 URL
          example: /api/uploads/id-card-front.jpg
          minLength: 1
        tenantIdCard:
          type: string
          description: 18 位居民身份证号
          example: "510100199001011234"
          minLength: 1
          pattern: "^\\d{17}[\\dXx]$"
        tenantName:
          type: string
          description: 租客真实姓名
          example: 张三
          maxLength: 30
          minLength: 1
        tenantPhone:
          type: string
          description: 租客联系电话
          example: "13800138000"
          minLength: 1
          pattern: "^1\\d{10}$"
      required:
      - idCardBackUrl
      - idCardFrontUrl
      - tenantIdCard
      - tenantName
      - tenantPhone
...
    RentOrderResponse:
      type: object
      description: 租房订单详情
      properties:
        address:
          type: string
          description: 详细地址
        cancelledAt:
          type: string
          format: date-time
          description: 订单取消时间
        contractConfirmedAt:
          type: string
          format: date-time
          description: 合同确认时间
        createdAt:
          type: string
          format: date-time
          description: 创建时间
        deposit:
          type: integer
          format: int32
          description: 押金，单位元
        endDate:
          type: string
          format: date
          description: 结束日期
        firstPaymentAmount:
          type: integer
          format: int32
          description: 首期应付金额，单位元
        houseId:
          type: string
          description: 房源 ID
        houseName:
          type: string
          description: 房源名称
        id:
          type: string
          description: 记录 ID
        leaseMonths:
          type: integer
          format: int32
          description: 租期月数
        monthlyRent:
          type: integer
          format: int32
          description: 月租金，单位元
        paidAt:
          type: string
          format: date-time
          description: 付款完成时间
        paymentMethod:
          type: string
          description: 租金付款方式
        paymentMonths:
          type: integer
          format: int32
          description: 每期支付月数
        realNameAt:
          type: string
          format: date-time
          description: 实名认证提交时间
        roomName:
          type: string
          description: 房间名称
        serviceFee:
          type: integer
          format: int32
          description: 服务费，单位元
        signedAt:
          type: string
          format: date-time
          description: 签约完成时间
        startDate:
          type: string
          format: date
          description: 开始日期
        status:
          type: string
          description: 业务状态
        tenantCount:
          type: integer
          format: int32
          description: 租住人数
        tenantIdCard:
          type: string
          description: 租客身份证号
        tenantName:
          type: string
          description: 租客姓名
        tenantPhone:
          type: string
          description: 租客联系电话
        totalAmount:
          type: integer
          format: int32
          description: 订单总金额，单位元
        updatedAt:
          type: string
          format: date-time
          description: 最后更新时间
        userId:
          type: string
          description: 用户 ID
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: "访问令牌，格式：Bearer {accessToken}"
      name: Authorization
      scheme: bearer
      type: http